Change doubles to floats in the PID library
Having a difference in type between the values given by the MPU
and those used by the PID cause issues in typecasting between
them.

While one solution would be to make a copy of the values in
the right type, the `double` precision offered by the PID was
unnecessary precision. So it seemed to be better to change
everything in the PID library to use floats to match what
is used by the MPU.