Motor: change PWM period from 20ms to 4ms
The period of the PWM signal was being set to 50Hz. This becomes a
bottleneck since the PID with the MPU run at 200Hz. The ArduPilot
project runs their motors with a frequency of 490Hz close to the limit
of 2ms. Since our maximum update rate is limited to 200Hz, a period of
4ms or a frequency of 250Hz should satisfy our requirements.

Upon testing the change on the motors, the motors seemed to operate
nominally.

Helped-By: Karthik Nayak <karthik.188@gmail.com>