Add centralized logger class
Motivation behind adding a centralized logger is:
* Significant stalls were noticed when writing out to stdout
  (redirected to file) of about 0.8s when the loop times were
  just 0.005s. This is a significant risk and take the
  quadcopter straight out of the air.

* Debugging messages were written from all over the code base
  based on a few flags which were set in different files as well.

This is the initial implementation of a centralized logger that
ensures that long lines of debug/log information is not written to
file at once or buffered to be written at once.

It aims to collect information once and write this data out chunk-
by-chunk. It does this by forming a log entry once and storing it.
It then writes out `freq` number of chunks of the log entry on
every call to update() until the log entry is completely written.
At this point, it will form a new log entry and repeat.

update() maintains a pointer to where it must continue on the
subsequent call to update().