Remove minimum delta_time between PID updates
`loop()` blocks while waiting for a value from the MPU6050. At current
rates, that is every 5ms. The only scenario when the `update_flag` is
false is when the time between loops is less than the minimum
delta time. Since we can guarantee a delay of at least 5ms, this
check, and hence `sample_time` is unnecessary