Controls: Add rate PID
A single PID is proving to be difficult to tune. After referring to
code from https://github.com/antodoms/beagle-copter and
https://github.com/cleanflight/cleanflight/blob/master/src/main/flight/pid_mwrewrite.c
a double PID seems like a better approach.

Beagle-copter uses an angle/stabilization PID feeding into a rate PID.
Desired rate of rotation is provided by the stab PID. (PID-PID)

Cleanflight, on the other hand, seems to be using just a P-PID
controller, where the rate PID's output is proportional to the
instantaneous angle error.

https://www.quora.com/What-is-rate-and-stabilize-PID-in-quadcopter-control/answer/Edouard-Leurent
shows how a PI-P controller is equivalent to a single PID controller.
This setup allows for easier tuning.

Although it is uncertain what benefits a PID-PID controller offers over
a single PID controller in practice, making tuning easier is a useful
addition.

This commit implements a PID-PID controller. To use it as a PI-P or a
P-PID type controller, setting those constants to 0 will work.

To implement the rate PID, we also utilize dmpGetGyro() which gives us
the current angular velocity about the X/Y/Z axis.

Helped-By: Karthik Nayak <karthik.188@gmail.com>