PID: Pass error calculation function pointer to constructor
Now that we are using two types of PIDs, each with a slightly
different error calculation requirement, either the error and delta
input must be passed in, or the function to calculate the difference
must be passed in.

This commit implements the latter method, allowing the user to
pass a pointer to the correct error calculation function.