MPU: Reduce sensitivity from 2000 to 250.
The MPU6050 when operated on its default value (2000) is really
sensitive. This essentially might make the Quad-copter unstable by
trying to minimize the minute differences in the orientation error
arising due to its high sensitivity. Based on this assumption and
reference from https://github.com/vjaunet/QUADCOPTER, change the
sensitivity to 250.

Signed-off-by: Karthik Nayak <Karthik.188@gmail.com>